rename script, remove old svg helper
