+++ src/partials/article.hbs
@@ -1,8 +1,7 @@
 <article class="doc">
   {{#if page.title}}
   {{#if page.attributes.icon}}
-  <h1 class="page"><i
-      class="border rounded-xl p-1{{#with page.attributes.icon-role}} {{{this}}}{{/with}}">{{{catalog-svg
+  <h1 class="page"><i class="border rounded-xl p-1{{#with page.attributes.icon-role}} {{{this}}}{{/with}}">{{{svg
       page.attributes.icon}}}</i>{{{page.title}}}</h1>
   {{else}}
   <h1 class="page">{{{page.title}}}</h1>
