+++ src/partials/tutorial-info.hbs
@@ -19,8 +19,8 @@
   {{/with}}
   <div class="tutorial-actions">
     {{#with page.attributes.colab-link}}
-    <a href="{{{this}}}" target="_blank" class="colab btn btn-outlined btn-neutral">
-      <div class="mr-1">{{{svg 'colab'}}}</div>
+    <a href="{{{this}}}" target="_blank" class="btn btn-outlined btn-neutral">
+      <i class="mr-1 colab-icon"></i>
       Run this on Colab
     </a>
     {{/with}}
